move some gulp file scripting to individual files
