+++ gulp/promisify-stream.js
@@ -1,0 +1,5 @@
+module.exports = function promisifyStream(stream) {
+	return new Promise((resolve, reject) => {
+		stream.on('finish', resolve).on('error', reject)
+	});
+};
